apps/api/pyproject.toml
[project]
name = "techbirds-api"
version = "0.0.0"
description = "TechBirds platform API (FastAPI + SQLAlchemy + Celery)."
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.2",
  "uvicorn[standard]==0.31.1",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "sqlalchemy==2.0.35",
  "asyncpg==0.29.0",
  "psycopg[binary]==3.2.3",
  "alembic==1.13.3",
  "redis==5.1.1",
  "celery[redis]==5.4.0",
  "passlib[bcrypt]==1.7.4",
  "python-jose[cryptography]==3.3.0",
  "httpx==0.27.2",
  "python-multipart==0.0.12",
  "boto3==1.35.36",
  "jinja2==3.1.4",
  "weasyprint==62.3",
  "sendgrid==6.11.0",
  "sentry-sdk[fastapi]==2.16.0",
  "structlog==24.4.0",
  "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
  "ruff==0.6.9",
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "pytest-cov==5.0.0",
  "mypy==1.11.2",
  "types-passlib==1.7.7.20240819",
]
prod = [
  "gunicorn==23.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "ASYNC"]
ignore = ["S101", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S"]
"alembic/**" = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
